Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/dom/events/Event-dispatch-throwing-expected.txt

CONSOLE MESSAGE: Error: Error from only listener
CONSOLE MESSAGE: Error: Error from first listener

PASS Throwing in event listener with a single listeners
PASS Throwing in event listener with multiple listeners
